Refactor tests with esm
Tap supports esm out of the box. We may just use it even with commonjs
implementation. The syntax is nice and no need to add `use strict`
annotation to each file.
