+++ test/get-codes.js
@@ -1,8 +1,6 @@
-'use strict'
-
-var { test } = require('tap')
-var { getCodes } = require('../')
-var data = require('../data.json')
+import { test } from 'tap'
+import { getCodes } from '../country-list.js'
+import data from '../data.json'
 
 test('get all country codes', function (t) {
   t.equal(getCodes().length, data.length, 'codes should be as many as countries')
